+++ src/data/navigaton-data.js
@@ -1,0 +1,27 @@
+
+export const inPlaceNavLinks = [
+    {
+        name: 'Bubble Sort',
+        url: ''
+    },
+
+    {
+        name: 'Selection Sort',
+        url: ''
+    },
+
+    {
+        name: 'Insertion Sort',
+        url: ''
+    },
+
+    {
+        name: 'Quick Sort',
+        url: ''
+    },
+
+    {
+        name: 'Heap Sort',
+        url: ''
+    },
+];