+++ src/index.js
@@ -1,16 +1,5 @@
-import Footer from './views/footer/Footer';
-import Header from './views/header/Header';
-import Main from './views/main/Main';
 import './styles/index.css';
+import App from './App';
 
-
-const body = document.body;
-
-const header = new Header();
-body.appendChild(header.create());
-
-const main = new Main();
-body.appendChild(main.create());
-
-const footer = new Footer()
-body.appendChild(footer.create());
+const app = new App()
+app.run();